Изменил алгоритм отрисовки ленты - теперь она собирается по чанкам. Исправил долгую загрузку страниц. Добавил на страницу регистрации коллаж с картинками
